Backend/requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1691763562544/work
asyncio==3.4.3
asyncpg @ file:///Users/runner/miniforge3/conda-bld/asyncpg_1700296520309/work
attrs==24.2.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.7.4
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
